requirements.txt
flask
flask-cors
langchain-ollama
langchain
langchain-community
langchain-core
langchain-openai
openai
pypdf
bs4
langgraph
ipython
jupyterlab
atlassian-python-api
lxml
chromadb
pdfminer.six
boto3
python-dotenv
jira